Move exceptions into the files that raise them
These exception classes are actually only ever raised (and only intended
to be raised) by these particular services, so move them into the same
files with the service classes that they belong to.
