+++ lms/services/launch_verifier.py
@@ -2,13 +2,24 @@
 from oauthlib.oauth1 import RequestValidator, SignatureOnlyEndpoint
 
 from lms import models
-from lms.services import (
-    ConsumerKeyLaunchVerificationError,
-    LTILaunchVerificationError,
-    LTIOAuthError,
-)
 
-__all__ = ["LaunchVerifier"]
+
+class LTILaunchVerificationError(Exception):
+    """
+    Raised when LTI launch request verification fails.
+
+    This is the base class for all LTI launch request verification errors.
+    Different subclasses of this exception class are raised for specific
+    failure types.
+    """
+
+
+class ConsumerKeyLaunchVerificationError(LTILaunchVerificationError):
+    """Raised when a given ``consumer_key`` doesn't exist in the DB."""
+
+
+class LTIOAuthError(LTILaunchVerificationError):
+    """Raised when OAuth signature verification of a launch request fails."""
 
 
 class LaunchVerifier:
